services/historical-data-service/requirements.txt
flask==2.3.3
gunicorn==21.2.0
pydantic==2.4.2
SQLAlchemy==2.0.20
asyncpg==0.28.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
structlog==23.1.0
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-instrumentation-flask==0.39b0
opentelemetry-instrumentation-sqlalchemy==0.39b0
opentelemetry-instrumentation-requests==0.39b0
kafka-python==2.0.2
pytest==7.4.0
pytest-asyncio==0.21.1
requests==2.31.0
httpx==0.24.1
pytest-cov==4.1.0
pyjwt==2.8.0
